src/background-job/requirements.txt
Flask
azure-storage-blob
azure.storage.queue